Fix dtype mismatch in generate_jit diffusion_steps
Fixed RuntimeError: mat1 and mat2 must have the same dtype error
by explicitly specifying dtype=torch.float32 when creating diffusion_steps
in the _sampling_label_jit method.

This ensures compatibility with the model's float32 weights and prevents
the default double (float64) type from being used.

+++ src/sssd/model_wrapper.py
@@ -270,7 +270,7 @@ class SSSDECG(nn.Module):
 
         # Reverse diffusion process
         for t in range(T-1, -1, -1):
-            diffusion_steps = (t * torch.ones((size[0], 1), device=self.device))
+            diffusion_steps = (t * torch.ones((size[0], 1), dtype=torch.float32, device=self.device))
 
             # Predict epsilon
             epsilon_theta = net((x, cond, diffusion_steps))
